+++ src/test/java/de/gematik/demis/notification/builder/demis/fhir/notification/builder/infectious/disease/NotificationBundleDiseaseDataBuilderTest.java
@@ -214,4 +214,54 @@ class NotificationBundleDiseaseDataBuilderTest {
       assertThat(excerptJson).isEqualTo(expected);
     }
   }
+
+  @Test
+  void shouldCreateExcerptWithGenderExtension() throws IOException {
+    int[] idHelper = {50};
+    try (MockedStatic<Utils> utils = Mockito.mockStatic(Utils.class)) {
+      utils
+          .when(Utils::generateUuidString)
+          .thenAnswer(invocation -> Integer.toString(idHelper[0]++));
+      utils.when(() -> Utils.getShortReferenceOrUrnUuid(any())).thenCallRealMethod();
+
+      String json =
+          Files.readString(
+              Path.of(
+                  "src/test/resources/disease/DiseaseNotificationTestcaseForNotByNameExcerpt_SpecialGenderCase.json"));
+      IParser iParser = FhirContext.forR4().newJsonParser().setPrettyPrint(true);
+      Bundle bundle = iParser.parseResource(Bundle.class, json);
+
+      Bundle excerpt = NotificationBundleDiseaseDataBuilder.createNonNominalExcerpt(bundle);
+      String excerptJson = iParser.encodeResourceToString(excerpt);
+
+      String expected =
+          Files.readString(
+              Path.of(
+                  "src/test/resources/disease/DiseaseNotificationTestcaseForNotByNameExcerpt_SpecialGenderCaseExpected.json"));
+
+      assertThat(excerptJson).isEqualTo(expected);
+    }
+  }
+
+  @Test
+  void shouldCreateExcerptWithOccuranceAsStringCase() throws IOException {
+    int[] idHelper = {50};
+    try (MockedStatic<Utils> utils = Mockito.mockStatic(Utils.class)) {
+      utils
+          .when(Utils::generateUuidString)
+          .thenAnswer(invocationOnMock -> Integer.toString(idHelper[0]++));
+      utils.when(() -> Utils.getShortReferenceOrUrnUuid(any())).thenCallRealMethod();
+
+      String json =
+          Files.readString(Path.of("src/test/resources/disease/occuranceCaseExcerpt.json"));
+      IParser iParser = FhirContext.forR4Cached().newJsonParser().setPrettyPrint(true);
+
+      Bundle bundle = iParser.parseResource(Bundle.class, json);
+
+      Bundle excerpt = NotificationBundleDiseaseDataBuilder.createNonNominalExcerpt(bundle);
+
+      assertThat(((Immunization) excerpt.getEntry().get(8).getResource()).getOccurrence())
+          .hasToString("ASKU");
+    }
+  }
 }